\name{kmeans.ddR-package}
\alias{kmeans.ddR-package}
\alias{kmeans.ddR}
\docType{package}
\title{Distributed clustering for Big Data}
\description{
\pkg{kmeans.ddR} provides a distributed training and predict functions for kmeans 
}

\details{
    \tabular{ll}{
        Package: \tab kmeans.ddR\cr
        Type: \tab Package\cr
        Version: \tab 0.1.0\cr
        Date: \tab 2015-10-28\cr
    }
 Main Functions:
 \itemize{
  \item {dkmeans:} {It is a distributed version of kmeans.}
 }
}

\author{
Vishrut Gupta, Arash Fard
}

\keyword{kmeans}
\keyword{clustering}
\keyword{distributed}
\keyword{distributed kmeans}
\keyword{distributed clustering}

